// src/app/models/product.ts
export interface Product {
  nameDisplay: string;
  nameLink: string;
  categoryDisplay: string;
  categoryLink: string;
  subCategoryDisplay: string;
  subCategoryLink: string;
  info: string;
  id: number;
  company: string;
  imgUrl: string;
  imgRef: string;
  price: number;
  quantity?: number;

}